281: create new /plan file structure
